LPS-84633 Extracting base test case
