+++ modules/apps/asset/asset-auto-tagger-test/src/testIntegration/java/com/liferay/asset/auto/tagger/internal/test/AssetAutoTaggerTest.java
@@ -15,50 +15,23 @@
 package com.liferay.asset.auto.tagger.internal.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.asset.auto.tagger.AssetAutoTagProvider;
 import com.liferay.asset.auto.tagger.AssetAutoTagger;
 import com.liferay.asset.auto.tagger.model.AssetAutoTaggerEntry;
 import com.liferay.asset.auto.tagger.service.AssetAutoTaggerEntryLocalService;
+import com.liferay.asset.auto.tagger.test.BaseAssetAutoTaggerTestCase;
 import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.asset.kernel.model.AssetTag;
-import com.liferay.asset.kernel.service.AssetEntryLocalService;
 import com.liferay.asset.kernel.service.AssetTagLocalService;
-import com.liferay.document.library.kernel.model.DLFileEntryConstants;
-import com.liferay.document.library.kernel.model.DLFolderConstants;
 import com.liferay.document.library.kernel.service.DLAppServiceUtil;
-import com.liferay.portal.configuration.test.util.ConfigurationTemporarySwapper;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
-import com.liferay.portal.kernel.test.util.GroupTestUtil;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.util.ArrayUtil;
-import com.liferay.portal.kernel.util.ContentTypes;
-import com.liferay.portal.kernel.util.HashMapDictionary;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.service.test.ServiceTestUtil;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.registry.Registry;
-import com.liferay.registry.RegistryUtil;
-import com.liferay.registry.ServiceRegistration;
 
-import java.util.Arrays;
-import java.util.Dictionary;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-import org.junit.After;
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
@@ -69,7 +42,7 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 @Sync
-public class AssetAutoTaggerTest {
+public class AssetAutoTaggerTest extends BaseAssetAutoTaggerTestCase {
 
 	@ClassRule
 	@Rule
@@ -78,38 +51,13 @@ public class AssetAutoTaggerTest {
 			new LiferayIntegrationTestRule(),
 			SynchronousDestinationTestRule.INSTANCE);
 
-	@Before
-	public void setUp() throws Exception {
-		ServiceTestUtil.setUser(TestPropsValues.getUser());
-
-		_group = GroupTestUtil.addGroup();
-
-		_serviceContext = ServiceContextTestUtil.getServiceContext(
-			_group.getGroupId(), 0);
-
-		Registry registry = RegistryUtil.getRegistry();
-
-		Map<String, Object> properties = new HashMap<>();
-
-		properties.put("model.class.name", DLFileEntryConstants.getClassName());
-
-		_assetAutoTagProviderServiceRegistration = registry.registerService(
-			AssetAutoTagProvider.class,
-			model -> Arrays.asList(_ASSET_TAG_NAME_AUTO), properties);
-	}
-
-	@After
-	public void tearDown() {
-		_assetAutoTagProviderServiceRegistration.unregister();
-	}
-
 	@Test
 	public void testAutoTagsANewAssetOnCreation() throws Exception {
-		_withAutoTaggerEnabled(
+		withAutoTaggerEnabled(
 			() -> {
-				AssetEntry assetEntry = _addFileEntryAssetEntry();
+				AssetEntry assetEntry = addFileEntryAssetEntry();
 
-				_assertContainsAssetTagName(assetEntry, _ASSET_TAG_NAME_AUTO);
+				assertContainsAssetTagName(assetEntry, ASSET_TAG_NAME_AUTO);
 			});
 	}
 
@@ -117,9 +65,9 @@ public class AssetAutoTaggerTest {
 	public void testDeletesAssetAutoTaggerEntriesWhenAssetIsDeleted()
 		throws Exception {
 
-		_withAutoTaggerEnabled(
+		withAutoTaggerEnabled(
 			() -> {
-				AssetEntry assetEntry = _addFileEntryAssetEntry();
+				AssetEntry assetEntry = addFileEntryAssetEntry();
 
 				List<AssetAutoTaggerEntry> assetAutoTaggerEntries =
 					_assetAutoTaggerEntryLocalService.getAssetAutoTaggerEntries(
@@ -143,7 +91,7 @@ public class AssetAutoTaggerTest {
 
 	@Test
 	public void testIsDisabledByDefault() throws Exception {
-		AssetEntry assetEntry = _addFileEntryAssetEntry();
+		AssetEntry assetEntry = addFileEntryAssetEntry();
 
 		List<AssetTag> assetTags = assetEntry.getTags();
 
@@ -152,19 +100,19 @@ public class AssetAutoTaggerTest {
 
 	@Test
 	public void testKeepsAssetTagCount() throws Exception {
-		_withAutoTaggerEnabled(
+		withAutoTaggerEnabled(
 			() -> {
-				AssetEntry assetEntry = _addFileEntryAssetEntry();
+				AssetEntry assetEntry = addFileEntryAssetEntry();
 
 				AssetTag assetTag = _assetTagLocalService.getTag(
-					_group.getGroupId(), _ASSET_TAG_NAME_AUTO);
+					group.getGroupId(), ASSET_TAG_NAME_AUTO);
 
 				Assert.assertEquals(1, assetTag.getAssetCount());
 
 				_assetAutoTagger.untag(assetEntry);
 
 				assetTag = _assetTagLocalService.getTag(
-					_group.getGroupId(), _ASSET_TAG_NAME_AUTO);
+					group.getGroupId(), ASSET_TAG_NAME_AUTO);
 
 				Assert.assertEquals(0, assetTag.getAssetCount());
 			});
@@ -172,13 +120,13 @@ public class AssetAutoTaggerTest {
 
 	@Test
 	public void testRemovesAutoTags() throws Exception {
-		_withAutoTaggerEnabled(
+		withAutoTaggerEnabled(
 			() -> {
-				AssetEntry assetEntry = _addFileEntryAssetEntry();
+				AssetEntry assetEntry = addFileEntryAssetEntry();
 
-				_assertContainsAssetTagName(assetEntry, _ASSET_TAG_NAME_AUTO);
+				assertContainsAssetTagName(assetEntry, ASSET_TAG_NAME_AUTO);
 
-				_applyAssetTagName(assetEntry, _ASSET_TAG_NAME_MANUAL);
+				applyAssetTagName(assetEntry, ASSET_TAG_NAME_MANUAL);
 
 				_assetAutoTagger.untag(assetEntry);
 
@@ -187,67 +135,9 @@ public class AssetAutoTaggerTest {
 				Assert.assertEquals(
 					assetTagNames.toString(), 1, assetTagNames.length);
 
-				_assertContainsAssetTagName(assetEntry, _ASSET_TAG_NAME_MANUAL);
+				assertContainsAssetTagName(assetEntry, ASSET_TAG_NAME_MANUAL);
 			});
 	}
-
-	private AssetEntry _addFileEntryAssetEntry() throws PortalException {
-		FileEntry fileEntry = DLAppServiceUtil.addFileEntry(
-			_group.getGroupId(), DLFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			RandomTestUtil.randomString(), ContentTypes.TEXT_PLAIN,
-			RandomTestUtil.randomString(), StringUtil.randomString(),
-			StringUtil.randomString(), new byte[0], _serviceContext);
-
-		return _assetEntryLocalService.getEntry(
-			DLFileEntryConstants.getClassName(), fileEntry.getFileEntryId());
-	}
-
-	private void _applyAssetTagName(AssetEntry assetEntry, String assetTagName)
-		throws PortalException {
-
-		_assetEntryLocalService.updateEntry(
-			assetEntry.getUserId(), assetEntry.getGroupId(),
-			assetEntry.getClassName(), assetEntry.getClassPK(),
-			assetEntry.getCategoryIds(),
-			ArrayUtil.append(assetEntry.getTagNames(), assetTagName));
-	}
-
-	private void _assertContainsAssetTagName(
-		AssetEntry assetEntry, String assetTagName) {
-
-		for (AssetTag assetTag : assetEntry.getTags()) {
-			if (StringUtil.equals(assetTag.getName(), assetTagName)) {
-				return;
-			}
-		}
-
-		throw new AssertionError(
-			String.format(
-				"The asset entry has not been tagged with '%s'", assetTagName));
-	}
-
-	private void _withAutoTaggerEnabled(UnsafeRunnable unsafeRunnable)
-		throws Exception {
-
-		Dictionary<String, Object> dictionary = new HashMapDictionary<>();
-
-		dictionary.put("enabled", true);
-
-		try (ConfigurationTemporarySwapper configurationTemporarySwapper =
-				new ConfigurationTemporarySwapper(
-					_CONFIGURATION_PID, dictionary)) {
-
-			unsafeRunnable.run();
-		}
-	}
-
-	private static final String _ASSET_TAG_NAME_AUTO = "auto tag";
-
-	private static final String _ASSET_TAG_NAME_MANUAL = "manual tag";
-
-	private static final String _CONFIGURATION_PID =
-		"com.liferay.asset.auto.tagger.internal.configuration." +
-			"AssetAutoTaggerSystemConfiguration";
 
 	@Inject
 	private AssetAutoTagger _assetAutoTagger;
@@ -255,25 +145,7 @@ public class AssetAutoTaggerTest {
 	@Inject
 	private AssetAutoTaggerEntryLocalService _assetAutoTaggerEntryLocalService;
 
-	private ServiceRegistration<AssetAutoTagProvider>
-		_assetAutoTagProviderServiceRegistration;
-
-	@Inject
-	private AssetEntryLocalService _assetEntryLocalService;
-
 	@Inject
 	private AssetTagLocalService _assetTagLocalService;
-
-	@DeleteAfterTestRun
-	private Group _group;
-
-	private ServiceContext _serviceContext;
-
-	@FunctionalInterface
-	private interface UnsafeRunnable {
-
-		public void run() throws Exception;
-
-	}
 
 }
